removes "no need to install packer" sentence
The next page (following the link) asks you if you have packer installed first
thing... and has a link to download packer if you say no... so it does not seem
like this sentence makes should be here.

While it is true there is no need to install packer before going to the tutorial
this sentence being here leads me to believe I will not need to install packer
at all for the tutorial which is not the case.